+++ Realm.Shared/Realm.cs
@@ -420,13 +420,11 @@ namespace Realms
             return CreateObject(className, out ignored);
         }
 
-        private RealmObject CreateObject(string className, out RealmObject.Metadata metadata)
+
+        internal RealmObject CreateObject(RealmObject.Metadata metadata)
         {
             if (!IsInTransaction)
                 throw new RealmOutsideTransactionException("Cannot create Realm object outside write transactions");
-
-            if (!Metadata.TryGetValue(className, out metadata))
-                throw new ArgumentException($"The class {className} is not in the limited set of classes for this realm");
 
             var result = metadata.Helper.CreateInstance();
 
@@ -435,6 +433,14 @@ namespace Realms
 
             result._Manage(this, rowHandle, metadata);
             return result;
+        }
+
+
+        private RealmObject CreateObject(string className, out RealmObject.Metadata metadata)
+        {
+            if (!Metadata.TryGetValue(className, out metadata))
+                throw new ArgumentException($"The class {className} is not in the limited set of classes for this realm");
+            return CreateObject(metadata);
         }